CASSANDRA-19500 Fix XXHash32Digest calculated digest value (#46)
This PR bumps the Sidecar version to the current latest HEAD of Sidecar. Bumping the
version surfaced an issue with the way we are producing digest strings for the XXHash32
implementation. The hash value is not masked and this causes the negative sign to be
forwarded producing the incorrect hash result.

Patch by Francisco Guerrero; Reviewed by Yifan Cai for CASSANDRA-19500

+++ cassandra-analytics-core/src/main/java/org/apache/cassandra/spark/utils/XXHash32DigestAlgorithm.java
@@ -61,7 +61,7 @@ public class XXHash32DigestAlgorithm implements DigestAlgorithm
             {
                 hasher.update(buffer, 0, len);
             }
-            return new XXHash32Digest(Long.toHexString(hasher.getValue()), SEED);
+            return new XXHash32Digest(Integer.toHexString(hasher.getValue()), SEED);
         }
     }
 }
